requirements/base.txt
Django>=1.8
django-environ>=0.3.0
django-braces>=1.8.0
django-crispy-forms>=1.4.0
django-admin-bootstrapped>=2.4.0
#django-authtools>=1.2.0
easy-thumbnails>=2.2
django-import-export>=0.2.7
django-simple-menu>=1.1.0
django-tinymce>=1.5.3
django-tables2>=1.0.4